install_k8s/gok-cloud/controller/backend/requirements.txt
Flask==3.0.3
flask-socketio==5.3.6
flask-jwt-extended==4.6.0
python-jose==3.3.0
requests==2.31.0
pika==1.3.2
werkzeug==3.0.1
PyYAML==6.0.1
# If you use Vault, add the client library you use, e.g.:
hvac==1.2.1